+++ test/blackbox-tests/gen_tests.ml
@@ -61,10 +61,11 @@ module Test = struct
     ; skip_platforms : Platform.t list
     ; enabled        : bool
     ; js             : bool
+    ; coq            : bool
     ; external_deps  : bool
     }
 
-  let make ?env ?skip_ocaml ?(skip_platforms=[]) ?(enabled=true) ?(js=false)
+  let make ?env ?skip_ocaml ?(skip_platforms=[]) ?(enabled=true) ?(js=false) ?(coq=false)
         ?(external_deps=false) name =
     { name
     ; env
@@ -73,6 +74,7 @@ module Test = struct
     ; external_deps
     ; enabled
     ; js
+    ; coq
     }
 
   let pp_sexp fmt t =
@@ -123,6 +125,7 @@ let exclusions =
   let odoc = make ~external_deps:true ~skip_ocaml:"4.02.3" in
   [ make "js_of_ocaml" ~external_deps:true ~js:true
       ~env:("NODE", Sexp.parse "%{bin:node}")
+  ; make "coq" ~external_deps:true ~coq:true
   ; make "github25" ~env:("OCAMLPATH", Dune_lang.atom "./findlib-packages")
   ; odoc "odoc"
   ; odoc "odoc-unique-mlds"
@@ -179,13 +182,14 @@ let pp_group fmt (name, tests) =
 
 let () =
   let tests = Lazy.force all_tests in
-  (* The runtest target has a "specoial" definition. It includes all tests
-     except for js and disabled tests *)
+  (* The runtest target has a "special" definition. It includes all
+     tests except for js, coq, and disabled tests *)
   tests |> List.iter ~f:(fun t -> Format.printf "%a@.@." Test.pp_sexp t);
-  [ "runtest", (fun (t : Test.t) -> not t.js && t.enabled)
+  [ "runtest", (fun (t : Test.t) -> not t.js && not t.coq && t.enabled)
   ; "runtest-no-deps", (fun (t : Test.t) -> not t.external_deps && t.enabled)
   ; "runtest-disabled", (fun (t : Test.t) -> not t.enabled)
-  ; "runtest-js", (fun (t : Test.t) -> t.js && t.enabled) ]
+  ; "runtest-js", (fun (t : Test.t) -> t.js && t.enabled)
+  ; "runtest-coq", (fun (t : Test.t) -> t.coq && t.enabled) ]
   |> List.map ~f:(fun (name, predicate) ->
     (name, List.filter tests ~f:predicate))
   |> Format.pp_print_list