+++ CHANGES.md
@@ -17,6 +17,8 @@ unreleased
 - Add experimental `$ dune init` command. This command is used to create or
   update project boilerplate. (#1448, fixes #159, @shonfeder)
 
+- Experimental Coq support (1466, @ejgallego)
+
 1.8.2 (10/03/2019)
 ------------------